rename auth2 to authentication
