+++ src/desktop/apps/authentication/__tests__/helpers.jest.js
