+++ src/desktop/apps/authentication/__tests__/reset_password.coffee
