+++ src/desktop/apps/authentication/client/__tests__/index.jest.js
