+++ src/desktop/apps/authentication/client/reset_password.coffee
