+++ src/desktop/apps/authentication/components/AuthStatic.tsx
