+++ src/desktop/apps/authentication/components/MobileAuthStatic.tsx
