+++ src/desktop/apps/authentication/components/__tests__/AuthStatic.jest.js
