+++ src/desktop/apps/authentication/components/__tests__/MobileAuthStatic.jest.js
