+++ src/desktop/apps/authentication/components/__tests__/ModalContainer.jest.js
