+++ src/desktop/apps/authentication/components/__tests__/meta.jest.js
