+++ src/desktop/apps/authentication/helpers.ts
