+++ src/desktop/apps/authentication/routes.ts
@@ -97,7 +97,7 @@ export const index = async (req, res, next) => {
       },
       locals: {
         ...res.locals,
-        assetPackage: "auth2",
+        assetPackage: "authentication",
       },
       data: {
         type,
