+++ src/desktop/apps/authentication/templates/reset_password.jade
