+++ src/desktop/apps/consign/components/create_account/index.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { connect } from "react-redux"
 import { Box } from "@artsy/palette"
 import { FormSwitcher } from "@artsy/reaction/dist/Components/Authentication/FormSwitcher"
-import { handleSubmit } from "desktop/apps/auth2/helpers"
+import { handleSubmit } from "desktop/apps/authentication/helpers"
 import { ModalHeader } from "reaction/Components/Modal/ModalHeader"
 import { updateAuthFormStateAndClearError } from "../../client/actions"
 import { ModalType } from "@artsy/reaction/dist/Components/Authentication/Types"
