--- src/desktop/assets/auth2.coffee
@@ -1,4 +1,0 @@
-$ = require('jquery')
-
-$ ->
-  require('../apps/auth2/client').init()
