+++ src/desktop/assets/authentication.coffee
@@ -1,0 +1,4 @@
+$ = require('jquery')
+
+$ ->
+  require('../apps/authentication/client').init()
