+++ src/desktop/assets/authentication.styl
@@ -1,0 +1,1 @@
+// This file was added so that the client-side js doesn't error out due to 404.
