+++ src/desktop/index.js
@@ -56,7 +56,7 @@ app.use(require("./apps/tag"))
 app.use(require("./apps/unsubscribe"))
 app.use(require("./apps/unsupported_browser"))
 app.use(require("./apps/style_guide"))
-app.use(require("./apps/auth2").app)
+app.use(require("./apps/authentication").app)
 app.use(require("./apps/static"))
 app.use(require("./apps/clear_cache"))
 app.use(require("./apps/sitemaps"))
